refactor(caches): rename files to match project conventions
- Rename keys.py to utils.py and sql.py to pg.py to align with project
  conventions (utils.py for utility functions, pg.py for PostgreSQL modules)
- Update all imports in data.py and tests accordingly
- Convert SQLCache rows to Cache model objects in get() and put() methods
